+++ packages/dialect-react-ui/components/ChatButton/index.tsx
@@ -25,9 +25,9 @@ function WrappedChatButton(
     refreshInterval: props.pollingInterval ?? DEFAULT_POLLING_INTERVAL,
   });
 
-  const wrapperRef = useRef<HTMLDivElement>(null);
-  const bellRef = useRef<HTMLButtonElement>(null);
-  useOutsideAlerter(wrapperRef, bellRef, close);
+  const refs = useRef<HTMLElement[]>([]);
+
+  useOutsideAlerter(refs, close);
 
   const { colors, bellButton, icons, modalWrapper, animations } = useTheme();
 
@@ -40,7 +40,10 @@ function WrappedChatButton(
     >
       <div className="dt-relative">
         <IconButton
-          ref={bellRef}
+          ref={(el) => {
+            if (!el) return;
+            refs.current[0] = el;
+          }}
           className={clsx(
             'dt-flex dt-items-center dt-justify-center dt-rounded-full focus:dt-outline-none dt-shadow-md',
             colors.bg,
@@ -60,7 +63,13 @@ function WrappedChatButton(
         show={ui?.open ?? false}
         {...animations.popup}
       >
-        <div ref={wrapperRef} className="dt-w-full dt-h-full">
+        <div
+          ref={(el) => {
+            if (!el) return;
+            refs.current[1] = el;
+          }}
+          className="dt-w-full dt-h-full"
+        >
           <Chat
             dialectId={props.dialectId}
             type="popup"