+++ packages/dialect-react-ui/components/common/constants.ts
@@ -1,0 +1,8 @@
+import type { Channel } from './types';
+
+export const DEFAULT_NOTIFICATIONS_CHANNELS = [
+  'web3',
+  'telegram',
+  'sms',
+  'email',
+] as Channel[];