+++ packages/dialect-react-ui/entities/wrappers/ConnectionWrapper.tsx
@@ -1,15 +1,56 @@
-import { useDialectConnectionInfo, useThreads } from '@dialectlabs/react-sdk';
+import {
+  useDialectConnectionInfo,
+  useDialectSdk,
+} from '@dialectlabs/react-sdk';
+import useSWR from 'swr';
 import NoConnectionError from '../errors/ui/NoConnectionError';
-import type { ReactNode } from 'react';
 
 // Only renders children if connected to successfully some backend
+type ConnectionValue = {
+  errorMessage: string;
+  isConnected: boolean;
+  isLoading: boolean;
+};
 interface ConnectionWrapperProps {
   header?: JSX.Element | null;
   pollingInterval?: number;
-  children?: ReactNode;
+  children?: JSX.Element | ((val: ConnectionValue) => JSX.Element | null);
 }
 
 // FIXME: trigger some hook to check connection
+
+interface UseDialectHealthProps {
+  baseUrl: string;
+  pollingInterval: number;
+}
+
+interface UseDialectHealthValue {
+  isOK: boolean;
+  isLoading: boolean;
+  error?: string;
+}
+
+// TODO: fix type
+const fetcher = (...args: any[]) => fetch(...args).then((res) => res.json());
+
+const useDialectHealth = ({
+  baseUrl,
+  pollingInterval,
+}: UseDialectHealthProps): UseDialectHealthValue => {
+  const { data, error } = useSWR(`${baseUrl}/api/v1/health`, fetcher, {
+    refreshInterval: pollingInterval,
+    revalidateOnReconnect: true,
+    refreshWhenHidden: true,
+  });
+
+  const isLoading = data === undefined && !error;
+
+  return {
+    isLoading,
+    error: error?.message,
+    isOK: data?.status === 'ok' && !error,
+  };
+};
 
 const DEFAULT_CONNECTIVITY_POLLING_INTERVAL = 5000;
 
@@ -19,7 +60,7 @@ export default function ConnectionWrapper({
   // potentially can be a separate setting completely, TBD.
   pollingInterval = DEFAULT_CONNECTIVITY_POLLING_INTERVAL,
   children,
-}: ConnectionWrapperProps): JSX.Element {
+}: ConnectionWrapperProps): JSX.Element | null {
   // TODO: take into account offline
   const {
     connected: {
@@ -34,20 +75,49 @@ export default function ConnectionWrapper({
     },
   } = useDialectConnectionInfo();
 
-  // FIXME: trigger some fetch from some of backends to get connection state
-  useThreads({
-    refreshInterval: pollingInterval,
+  const {
+    info: {
+      config: {
+        dialectCloud: { url: baseUrl },
+      },
+    },
+  } = useDialectSdk();
+
+  const { isOK, error, isLoading } = useDialectHealth({
+    baseUrl,
+    pollingInterval,
   });
 
-  const someBackendConnected =
+  const isSomeBackendConnected =
     (isSolanaShouldConnect && isSolanaConnected) ||
     (isDialectCloudShouldConnect && isDialectCloudConnected);
 
-  if (!someBackendConnected) {
+  const isConnected = isSomeBackendConnected && isOK;
+
+  const connectingTo = [
+    isDialectCloudShouldConnect &&
+      (!isDialectCloudConnected || !isOK) &&
+      'Dialect Cloud',
+    isSolanaShouldConnect && !isSomeBackendConnected && 'Solana',
+  ]
+    .filter(Boolean)
+    .join(' and ');
+
+  const errorMessage = `Error connecting to ${connectingTo}`;
+
+  if (typeof children === 'function') {
+    return children({
+      errorMessage: error ? errorMessage : '',
+      isConnected,
+      isLoading,
+    });
+  }
+
+  if (!isConnected && !isLoading) {
     return (
       <>
         {header}
-        <NoConnectionError />
+        <NoConnectionError message={errorMessage} />
       </>
     );
   }