+++ packages/dialect-react-ui/utils/useOutsideAlerter.ts
@@ -1,8 +1,7 @@
 import { RefObject, useEffect } from 'react';
 
 export function useOutsideAlerter(
-  wrapperRef: RefObject<HTMLDivElement>,
-  buttonRef: RefObject<HTMLButtonElement>,
+  refs: RefObject<HTMLElement[]>,
   close: CallableFunction
 ) {
   useEffect(() => {
@@ -10,14 +9,20 @@ export function useOutsideAlerter(
      * Alert if clicked on outside of element
      */
     function handleClickOutside(event: MouseEvent) {
-      if (
-        wrapperRef.current &&
-        !wrapperRef.current.contains(event.target as Node) &&
-        buttonRef.current &&
-        !buttonRef.current.contains(event.target as Node)
-      ) {
-        close();
+      if (!refs?.current) {
+        return;
       }
+
+      const clickWithinActiveElements = refs.current
+        .filter(Boolean)
+        .some((el) => el?.contains(event.target as Node));
+
+      if (clickWithinActiveElements) {
+        return;
+      }
+
+      // trigger close if click is outside desired elements
+      close();
     }
 
     // Bind the event listener
@@ -26,5 +31,5 @@ export function useOutsideAlerter(
       // Unbind the event listener on clean up
       document.removeEventListener('mousedown', handleClickOutside);
     };
-  }, [wrapperRef, buttonRef, close]);
+  }, [refs, close]);
 }